fix: Move docker builds to native platforms
The current Docker builds are taking an excessively long time to complete,
currently over 6 hours causing GitHub Actions to time out. We believe
it's because internally it's using virtualization to build the arm64
version of the container.

To attempt to alleviate this, this PR uses a GitHub Actions matrix to
separate the builds into runners on their own native platforms.

The hope is that the lack of needing to virtualize will speed up the
overall time.