chore: Turn on latest tagging for docker builds (#23)
Previously we turned this off to test that the multi-arch builds
actually work. We've verified that we were successfully able to build
and push a mulit-arch image, so now we can turn back on the latest
tagging.

This new approach changes things just a bit. It turns this from relying on
new tag pushes to instead being manually activated. Deploying now
will involved someone activating the github action to build the new
docker container and being able to choose whether it tags the image
it will build as "latest" or not.