chore: Remove branch tagging from docker builds
Previously I added the branch tagging because it seemed like it could
be useful for building from PRs, but since we've changed the way
the build works it ends up pushing a `master` tag now.

This tag can be confusing because it's unlikely to be up to date with
actual master, so it is being removed.