fix: Allow explicit input of docker image build tag
The fix introduced in previous patches did not carry over an important
piece of functionality in that it just assumed the docker image tag
was always going to align with the build version for the software.

This is not always so as the build tag might need to be incremented
due to changes being made on the docker side and not the software side.

This PR moves to instead of gleaning those values from the compose file
instead making the runner explicitly define them so that it's easier
to grok what the final values will be.