fix: Give docker build tags explicit vars
The previous changes to docker builds did not carry over the ability
to differentiate the docker image build tags from the version of the
software being build. This is important due to the fact that sometimes
we'll want to rebuild the docker image because of docker specific changes
but we don't want to force push a new digest over an already established
docker tag.

This PR moves `build_tag` and `build_version` to explicitly defined
variables so that it's a bit easier to grok where these values are
coming from and where they are being inserted.